Move Trie, Packages and Jar to 'base'
