+++ intellij-plugin-structure/structure-classes/src/main/java/com/jetbrains/plugin/structure/classes/resolvers/DirectoryResolver.kt
@@ -13,6 +13,7 @@ import com.jetbrains.plugin.structure.base.utils.toSystemIndependentName
 import com.jetbrains.plugin.structure.classes.utils.AsmUtil
 import com.jetbrains.plugin.structure.classes.utils.getBundleBaseName
 import com.jetbrains.plugin.structure.classes.utils.getBundleNameByBundlePath
+import com.jetbrains.plugin.structure.jar.Packages
 import org.objectweb.asm.tree.ClassNode
 import java.nio.file.Files
 import java.nio.file.Path
