+++ intellij-plugin-structure/structure-classes/src/main/java/com/jetbrains/plugin/structure/classes/resolvers/FixedClassesResolver.kt
@@ -6,6 +6,7 @@ package com.jetbrains.plugin.structure.classes.resolvers
 
 import com.jetbrains.plugin.structure.base.BinaryClassName
 import com.jetbrains.plugin.structure.classes.utils.getBundleBaseName
+import com.jetbrains.plugin.structure.jar.Packages
 import org.objectweb.asm.tree.ClassNode
 import java.util.*
