+++ intellij-plugin-structure/tests/src/test/kotlin/com/jetbrains/plugin/structure/classes/resolvers/jar/JarFixtures.kt
@@ -2,6 +2,7 @@ package com.jetbrains.plugin.structure.classes.resolvers.jar
 
 import com.jetbrains.plugin.structure.base.utils.contentBuilder.buildZipFile
 import com.jetbrains.plugin.structure.base.utils.emptyClass
+import com.jetbrains.plugin.structure.jar.Jar
 import com.jetbrains.plugin.structure.jar.JarFileSystemProvider
 import net.bytebuddy.ByteBuddy
 import java.nio.file.Path
