+++ intellij-plugin-structure/tests/src/test/kotlin/com/jetbrains/plugin/structure/classes/resolvers/jar/JarTest.kt
@@ -2,7 +2,10 @@ package com.jetbrains.plugin.structure.classes.resolvers.jar
 
 import com.jetbrains.plugin.structure.base.utils.contentBuilder.buildZipFile
 import com.jetbrains.plugin.structure.base.utils.emptyClass
+import com.jetbrains.plugin.structure.jar.Jar
+import com.jetbrains.plugin.structure.jar.JarEntryResolver
 import com.jetbrains.plugin.structure.jar.SingletonCachingJarFileSystemProvider
+import com.jetbrains.plugin.structure.jar.descriptors.DescriptorReference
 import net.bytebuddy.ByteBuddy
 import org.junit.Assert.*
 import org.junit.Before
